Se modifico db.json

+++ src/components/ItemCarrito.jsx
@@ -1,36 +1,44 @@
 import { useContext } from "react";
-import ProductosContext from "../contexts/productosContext";
 import CarritoContext from "../contexts/CarritoContext";
+import Swal from "sweetalert2";
 import "./ItemCarrito.scss";
 
 const ItemCarrito = ({ producto }) => {
   const { eliminarProductodelCarritoContext } = useContext(CarritoContext);
 
   const handleEliminar = (id) => {
-    eliminarProductodelCarritoContext(id);
-  };
-
-  const handleEditar = (producto) => {
-    console.log("Editar producto", producto);
-  };
+    Swal.fire({
+      title: "Estás seguro?",
+      text: "No podrás revertirlo!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "rgb(34, 204, 82)",
+      cancelButtonColor: "rgb(10, 76, 42)",
+      confirmButtonText: "Sí, borrar del carrito!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        eliminarProductodelCarritoContext(id);
+        Swal.fire({
+          title: "¡Eliminado!",
+          text: "El producto fue eliminado.",
+          icon: "success",
+        });
+      }
+    });
+  }; 
 
   return (
     <tr>
-      <td>{producto.nombre}</td>
-      <td>{producto.precio}</td>
-      <td>{producto.stock}</td>
-      <td>{producto.marca}</td>
-      <td>{producto.categoria}</td>
-      <td>{producto.detalle}</td>
-      <td>
+        <td>
         <img src={producto.foto} alt={producto.nombre} style={{ width: "60px" }} />
       </td>
-      <td>{producto.envio ? "si" : "no"}</td>
+      <td>{producto.nombre}</td>
+      <td>{producto.cantidad}</td>
+      <td>{producto.precio}</td>
       <td>
-        <button className="btn btn-ver">Ver</button>
         <button className="btn btn-borrar" onClick={() => handleEliminar(producto.id)}>Borrar</button>
-        <button className="btn btn-editar" onClick={() => handleEditar(producto)}>Editar</button>
       </td>
+      <td>{producto.cantidad * producto.precio}</td>
     </tr>
   );
 };
